templated jet probability tag computers
